server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai==1.12.0
google-generativeai==0.3.2
pydub==0.25.1
pyaudioop==0.3.5
python-multipart==0.0.6
